Add Swift Package Manager manifest file.
As of this commit, we can build under Swift 3 pre-release (2016-03-24 edition), in addition to Swift 2.x with Xcode 7.3. However, that's only true because there are only two (very simple) source files added to the SDK so far. Swift 3 is already wildly source-incompatible with Swift 2, so we will need a dedicated Swift 3 branch pretty soon.