Fix fundamentally bogus[1] design of the namespace feature of SoracomCredentials.
[1]: It was poor design when every client app had to set up a unique UUID for
credentials storage. That is fixed, and NSBundle's bundleIdentifier is now one
component of the unique storage key.

A 'namespace' is a conceptually independent storage area within one app, not
within the system as a whole. Therefore, the defaults should be (and now are)
useable without the client app having to take any action to make it so.

This includes unit tests, which were broken with the previous implementation
(they worked in one app, but once you had two apps using the SDK and tests
storing credentials, they were clobbering each other).

This is very similar to the problem in KeychainTests that was fixed recently in
9afee5a063b09f8a0a7ceafd221d10abacf24d0f.