Change Request's completion handlers to block and wait for completion.
This is in preparation for adding NSOperation support. This only blocks the subthread
processing the HTTP response, so I don't think it's a problem, and it makes things
easier to think about when adding NSOperationQueue with all its threading issues,
and async NSOperation's weird deadlock-prone KVO sensitivity, etc., into the mix.