Rename TestRunnerAppForSoracomSDK to MacDemoAppForSoracomSDK.
And update all the bits inside the Xcode project. This change reflects an expanded
role for this app; in addition to just running the tests, it now has the additional jobs
of making it easy for developers to play with the SDK, while also serving as the main
place where the SDK itself is developed.