Implement deferred Request construction in APIOperation.
This allows you to queue up multiple APIOperation instances,
some of which may depend upon data retrieved from the server
by earlier operations. The actual API request is not constructed
until the operation begins execution, so any data resulting from
earlier operations can be made available for constructing the
request.

A simple example of this would be a queue of operations where
an earlier request retrieves a token from the API server, and then
subsequent requests use that token to authorize some action.