Remove the NSOperationQueue extension[1], and update demo app[2].
[1]: With the recently-added ability for APIOperation to defer creation of
its Request object, the extension to NSOperationQueue didn't make as
much sense. (Actually, it was probably already not worth existing; it saved
a little bit of typing but did not make things clearer.)

[2]: Update the demo app to fix where it had used the no-longer-available
NSOperationQueue extension, and also update it to use the new "deferred"
RequestBuilder variant of APIOperation, where there was a good example
of needing that.