Simplify APIOperation, by making Request and ResponseHandler more flexible.
You can now supply a response handler for a Request three different ways:

1. supply it during initialize
2. just set it later
3. supply it as an argument to run()

That's 3 different possibilities to have to consider, but it enabled an overall simplification
for dealing with APIOperation and deferred requests. There was a code smell that bugged
me with APIOperation where you had to supply separate request builders and response
handlers to it, which was sort of the same but a little different from the way it was done
with Request.

Now, APIOperation doesn't have a response handler at all, and you just initialize one either
with a request instance (simple) or, for the deferred-request-construction case, a request
builder (simple).

And everything else then works exactly the same whether you are running a Request
manually, wrapping an existing request in an APIOperation, or queuing an operation
that will defer creation of the actual request until later.