Add iOS demo app.
This is similar to the Mac demo app, and is mostly intended to prove that the
SDK compiles for iOS also, and all the tests pass on that platform. However,
currently there are some issues to fix wrt the way credentials are stored and
retrieved, and this causes the iOS app to skip some of the more complex
tests. (See https://github.com/soracom/soracom-sdk-swift/issues/1) #1