Disambiguate naming, and fix bug[1] where demo app didn't save credentials for tests.
[1]: Bug was that the  demo app no longer saved credentials, when creating a sandbox user,
in a way that the unit tests would be able to read them and use them to run the tests. This
was annoying because it meant you would have to manually save credentials (e.g. in code or
in the debugger) every week or so, because they expire.

Bug was here:

https://github.com/soracom/soracom-sdk-swift/blob/0f199eb038b7fa00256f9fa1da40425339b55cd3/Whatever/MacDemoAppForSoracomSDK/MacDemoAppForSoracomSDK/AppDelegate%2BAPIExamples.swift#L186

... replaceDefault used to be true by default. When I changed that, I introduced this bug. Since
the fix is literally on the same line as some of the renaming in this commit, I couldn't really break
this fix out into its own commit.