Move definition of default credentials storage namespaces to SoracomCredentials.[1]
[1]: Initially I thought that only the tests needed to access these constants, but in the
real world, each app that uses them will probably want to be able to access them too,
if only to make it more convenient for the developers to enter credentials that allow
running tests against the API sandbox. This change just makes the constants available
to both the automated tests and the app hosting those tests. (This is specific to Xcode
and XCTest.)