Fix bug where auto-created API Sandbox user had no credit card registered.
Actually the bug was: after registering a payment method for the sandbox
user, and updating the API token, we didn't properly save the updated token.
This meant a stale token would be used, which caused the API server to
reject SIM registration with a "Payment method is not registered" error.

Despite being only three letters of code to fix, this bug took me 2 hours to
find... :puke: