Consolidate redundant and similar code into new Client class.
This class implements various steps that the Mac demo app, the
iOS demo app, and the automated tests previously did in a
chaotic and confusing manner.

This also standardizes how all of the demo apps read/write
credentials, and allows the iOS app to finally run all the tests as
well.