Refactor Payload, fixing the major issues with its design.
These changes go a long way towards making it clearer who
is responsible for [de]serialization at which level. Response
should not be involved in that, and no longer is; it's all
delegated to Payload.

Payload can now handle array root objects. My initial version
of this was a terrible kludge; since the backing store for the
payload's data was a dictionary, in the array case I made it
so that the array was stored there under the key "rootObject".

That sucked on multiple levels, but now the foundation is there
for having whatever different types of root object backing store
we need. Currently, that's just array and dictionary, but with the
new rootObjectType property keeping track of whether it is a
"foreign" object (types unknown, just a JSON structure) or a
"native" object (which might contain various rich SDK object
types, that Payload can serialize to JSON).

The toDictionary() method now returns an optional value, since
returning nil makes sense for a non-dictionary payload.

I think that these changes make Payload ready to handle the
rest of the Soracom API.

There is a little more work to do: I would like to make it also
generic enough to just handle arbitrary JSON going both ways,
and make all the key validation stuff completely opt-in.

The Response struct was also improved in this cleanup:

- better error check API (computed on demand now)

- removed getErrorIfExpectedKeysAreMissing because that
  is stupid; maybe a validation closure that would be invoked
  at the end of the network transaction could work better?