[WIP] Add tool to generate code with swagger.
This tool will obsolete the ./Whatever/swagger
dir and everything in it. The primordial script
that still exists there cannot work under Swift 3
due to the enforcement of the deprecated system()
call being unavailable in Swift 3 on. So I will just
trash that whole dir once this new tool is done.

Secondly, I realize this can't be a command-line
tool, it needs to be an app bundle so it can
contain resources (namely the JSON API
definition and our custom templates). That does
not mean it can't be run on the command line, it
just means we have to change the build.

(I'm committing this partial work because I will
move to Soracom HQ, but this does not yet fully
work.)