Map all swagger "integer" types to Swift Int64.[1]
[1]: I would actually love to eliminate 32-bit
support, which would mean that no mapping would
be required, as Int itself would be 64-bit on
the Swift side. But, we cannot do that, because
iPhone 5c and some old-model iPads will run iOS 10,
so that would mean those devices couldn't run the
upcoming iOS app (which is based on this SDK).

Also, who knows what weird little IoT devices might
emerge that use 32-bit architectures (and run Swift
code on Linux?? well, it's maybe possible...)

So we maintain 32-bit/64-bit compatibility, but as
a convention only ever use 64-bit integers (rather
than try to figure out when we need them and when
we don't...)