BOOM! 💣 Get the first Swagger-generated code[1] building and passing tests. 😁
This is the first milestone in the process toward replacing a lot of the
handwritten code with swagger-generated code, and adding implementations
for all of the remaining model objects.

[1]: I started with just Subscriber and its dependencies. The other generated
code is added to the project tree in this commit, but is not yet added to
the build. I expect there are still many problems to be worked out, but
need to start with a subset of the model objects so make it manageable.