Fix bug in Payload decode for Swift types not auto-bridged[1] to ObjC types.
[1]: Int type is auto-bridged, so a value of NSNumber can be "as? Int"
casted, but that doesn't work for types like Int64. This meant decode
for Int64 was not bi-directional; it would fail for source objects which
were decoded via NSJSONSerialization (because that decodes to ObjC objects,
so numbers are NSNumber instances). That, in turn, meant that the getInt()
and getInt64() methods would only work for *outgoing* JSON payloads, but
not incoming ones.

Trying the cast both ways fixes this, and we will have to do this for all
value types other than Apple's blessed magic types.