Update for Swift 3.0.
Note: All the development over the past month or so has been in the
experimental add-swagger-codegen branch, and so that is where things were
updated for compatibility with the various Swift 3 pre-releases. But now
that Swift 3 is actually out, the required changes have been pulled into
master.

(It's not 100% clear yet whether or not the add-swagger-codegen branch
will ultimately work out, or not.)