🎉 First successful[1] Linux build!
[1]: successful *build*; doesn't imply any working functionality yet...

Use URLRequest (not NSMutableURLRequest) on Linux, and fix name conflict.

The name conflict: Request has a property called URLRequest, but since
then Swift has added a type with that same name. We should rename the
property, but for now I just create the URLRequest using the fully-
qualified Foundation.URLRequest(...)