Create Xcode project for the package (and remove it from .gitignore)
Normally, when using Swift Package Manager, you want to generate the
.xcodeproj project file with 'swift package generate-xcodeproj' and
not track it in git. The idea is, to recreate it as necessary.

However, in this case there are some manual edits we have to do to the
project to get it to build. So for now, I am adding it to .git. We may
revisit this later and revert to the normal SPM way of doing things
(especially as Swift 4 and SPM mature, things might improve).