Invoke doInitialHousekeeping() early on in tests, so credentials may be entered.
This used to be done by the macOS demo app. Well, it still is, but now
we can run the SDK's tests *without* the macOS demo app. So, we needed
a way to do that same thing, so that credentials can be supplied for
the tests to use when running against the API Sandbox.