Divide storage namespacing roles between Keychain and SoracomCredentials
SoracomCredentials used to be responsible for ensuring globally unique
storage keys were used to save credentials. However, this was not the
right design; it just happened.

Keychain writes to a shared store (at least on some platforms), and
therefore the Keychain API should bear the responsibility of making
sure globally unique keys are used — typically there would only be one
unique storage identifier per app. This is now Keychain's job.

SoracomCredentials, on the other hand, does (sometimes) need unique
namespaces by which credentials are categorized. However, these need
only be unique within the scope of the app. They don't need to be
globally unique. So with this commit, they no longer are.

The roles of Keychain and SoracomCredentials are now more properly
scoped, and are hopefully more easily understood at a glance.