🚧 Begin moving to Codable: make Credential serialize via Codable
This is part of a multi-step process, to remove all NSObject- and
objc-based hacks for JSON serialization, and use the new Codable set
of protocols.

I want to do it piece by piece, so the Old Gods and the New Gods will
have to live together for a while.

In this commit, we use Credential as a test case for implementing new
serialization.

- not removing Payload for now
- but moving to Codable as method of serialization to/from Payload
- extend Encodable and Decodable to create base serialization mechanism
- make Payload → JSON work for more cases (temporary fix until removal)
- remove 90% of Credential's code, because Codable replaces it