Complete rewrite of Payload (drop-in replacement)
This is a total rewrite of Payload that replaces the original
implementation, and exposes the same API. This doesn't include all the
improvements I want to make, as those are best handled one at a time,
now that we have the new Payload passing all existing tests.

The code I want to revisit on a piece-by-piece basis is still there,
but organized into two extensions to Payload (labeled "dubious code").

INCLUDED:

- Encoding to JSON now always uses Swift 4 Encodable (via a pair of
  extensions to KeyedEncodingContainer & UnkeyedEncodingContainer

- There's no more weird dichotomy between "native" and "foreign" payloads

- Internal representation of root object is now just [Any] or
  [String:Any]. The mutability API still uses PayloadKey like
  foo[.apiKey] = "bar" but that doesn't carry over into internal data
  structures

- Our weird old friend coerceValueToBasicType() is gone (though his
  spirit lives on in our Encodable extensions)

NOT INCLUDED:

- Removal of intermediate dictionary/array representations: There's a
  bunch of code (mostly/all test code) that relies on going to/from
  array or dictionary of Swift primitives. This doesn't make sense for
  an object like Payload that is inherently expected to go to/from
  JSON

- well, lots of stuff; refer to "DUBIOUS CODE" I & II