Use an .xcconfig file to specify MACOSX_DEPLOYMENT_TARGET
This doesn't solve the problem that "swift package generate-xcodeproj"
uses 10.10 for this setting, and so macOS builds don't work without
user intervention to change the setting.

I think it would be better if there were a way to tell SPM to use the
.xcconfig file, though — either merely because it exists (easy), or
somehow specified in the Package.swift file (and enhancement that
would probably be more convoluted to get approved).

As it stands, SPM does *not* understand this, so it doesn't really
solve anything, beyond making explicit what the Xcode project settings
need to be. The user still has to take explicit action to get the Mac
build working, if we do in fact delete the .xcodeproj file from git...