Add basic command-line demo program (formerly "Linux demo app")
The command-line demo is basic, having just a few predefined functions.
It uses Request.wait() for simplicity, so it offers an alternative
example to Client, which is used in the macOS and iOS demo apps, and
tends to do things asynchronously.

This could be improved but it suffices for a basic Linux demo.