Proof-of-concept with autogenerated code via SwagGen
Replaced Subscriber, and the classes it depends on, with versions that
were generated with SwagGen. It seems to work well, although I had to
enable safeArrayDecoding and safeOptionalDecoding to get it working
with the existing code. This still leaves the stupid Payload class in
place from the pre-Codable days, but... leave that for later.