😓 First working BUILD of new swaggen-based generated code version[1]
[1] This is a huge break-the-world kind of commit. It's hacky and the
first time I got this to successfully build. Many things don't work,
though. This is the start of using autogenerated, Swift 4 Codable based,
model and request code to power the project. This will let us remove
Payload at last. BUT, some of the basic design of the SDK had to change
and things are still very messy.