Refactor the macOS demo app in a better[1] style
Initially we were just adding the SDK source to the demo app's Xcode
project. However, it is better to set it up so that the SDK project is
a subproject of the demo app, and Xcode builds the subproject and
depends on it automatically.

This cleans up the test importing code and is probably more reflective
of how real-world Xcode projects using this SDK would be configured.